services/vision-preprocess-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
transformers==4.57.0
torch==2.5.1
torchvision==0.20.1
accelerate==1.2.1
numpy==1.26.4
Pillow==11.3.0
python-multipart==0.0.20
requests==2.32.5
safetensors==0.5.3
timm==1.0.19
einops==0.8.1
